+++ test/runTest.js
@@ -1,5 +1,4 @@
 const path = require("path");
-
 const { runTests } = require("vscode-test");
 
 async function main() {